# src/WeatherApp.Api/generated/controllers/WeatherForecastsController.cs
// Generated by @typespec/http-server-csharp
// <auto-generated />
#nullable enable

using System;
using System.Net;
using System.Threading.Tasks;
using System.Text.Json;
using System.Text.Json.Serialization;
using Microsoft.AspNetCore.Mvc;
using WeatherApi;

namespace WeatherApi.Controllers
{
    [ApiController]
    public partial class WeatherForecastsController : ControllerBase
    {

        public WeatherForecastsController(IWeatherForecasts operations)
        {
            WeatherForecastsImpl = operations;
        }
        internal virtual IWeatherForecasts WeatherForecastsImpl { get; }

        ///<summary>
        /// Get all weather forecasts
        ///</summary>
        [HttpGet]
        [Route("/weatherforecast")]
        [ProducesResponseType((int)HttpStatusCode.OK, Type = typeof(WeatherForecast[]))]
        public virtual async Task<IActionResult> ListForecasts()
        {
            var result = await WeatherForecastsImpl.ListForecastsAsync();
            return Ok(result);
        }

        ///<summary>
        /// Add a new weather forecast
        ///</summary>
        [HttpPost]
        [Route("/weatherforecast")]
        [ProducesResponseType((int)HttpStatusCode.Created, Type = typeof(WeatherForecast))]
        public virtual async Task<IActionResult> AddForecast(WeatherForecast body)
        {
            var result = await WeatherForecastsImpl.AddForecastAsync(body);
            return Ok(result);
        }

    }
}