# src/WeatherApp.Api/generated/models/BadRequestError.cs
// Generated by @typespec/http-server-csharp
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;
using TypeSpec.Helpers.JsonConverters;
using TypeSpec.Helpers;

namespace WeatherApi
{

    ///<summary>
    /// Bad request error
    ///</summary>
    public partial class BadRequestError : ErrorResponse
    {
        public BadRequestError(string code = "BAD_REQUEST") : base(400,
           value: new { code = code })
        {
            Code = code;
        }
        public new string Code { get; } = "BAD_REQUEST";


    }
}