# src/WeatherApp.Api/generated/models/ErrorResponse.cs
// Generated by @typespec/http-server-csharp
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;
using TypeSpec.Helpers.JsonConverters;
using TypeSpec.Helpers;

namespace WeatherApi
{

    ///<summary>
    /// Standard error response
    ///</summary>
    public partial class ErrorResponse : HttpServiceException
    {
        public ErrorResponse(string code, string message, DateTimeOffset timestamp, string details = default, string requestId = default) : base(400,
           value: new { code = code, message = message, timestamp = timestamp, details = details, requestId = requestId })
        {
            Code = code;
            MessageProp = message;
            Timestamp = timestamp;
            Details = details;
            RequestId = requestId;
        }
        public ErrorResponse(int statusCode, object? value = null, Dictionary<string, string>? headers = default) : base(statusCode, value, headers) { }

        ///<summary>
        /// Error code
        ///</summary>
        public string Code { get; set; }

        ///<summary>
        /// Human-readable error message
        ///</summary>
        [JsonPropertyName("message")]
        public string MessageProp { get; set; }

        ///<summary>
        /// Additional error details
        ///</summary>
        public string Details { get; set; }

        ///<summary>
        /// Timestamp when error occurred
        ///</summary>
        public DateTimeOffset Timestamp { get; set; }

        ///<summary>
        /// Request ID for tracking
        ///</summary>
        public string RequestId { get; set; }


    }
}