# src/WeatherApp.Api/generated/models/WeatherForecast.cs
// Generated by @typespec/http-server-csharp
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;
using TypeSpec.Helpers.JsonConverters;
using TypeSpec.Helpers;

namespace WeatherApi
{

    ///<summary>
    /// Represents a weather forecast for a specific location and date
    ///</summary>
    public partial class WeatherForecast
    {
        ///<summary>
        /// Unique identifier for the weather forecast
        ///</summary>
        public string Id { get; set; }

        ///<summary>
        /// Date of the weather forecast
        ///</summary>
        public DateTime Date { get; set; }

        ///<summary>
        /// Temperature in Celsius
        ///</summary>
        public int TemperatureC { get; set; }

        ///<summary>
        /// Temperature in Fahrenheit (calculated)
        ///</summary>
        public int TemperatureF { get; set; }

        ///<summary>
        /// Weather summary description
        ///</summary>
        public string Summary { get; set; }

        ///<summary>
        /// Geographic location of the forecast
        ///</summary>
        public string Location { get; set; }


    }
}