# src/WeatherApp.Api/generated/models/WeatherForecastList.cs
// Generated by @typespec/http-server-csharp
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;
using TypeSpec.Helpers.JsonConverters;
using TypeSpec.Helpers;

namespace WeatherApi
{

    ///<summary>
    /// List of weather forecasts
    ///</summary>
    public partial class WeatherForecastList
    {
        ///<summary>
        /// Array of weather forecast items
        ///</summary>
        public WeatherForecast[] Items { get; set; }


    }
}